Splits `kda.core` into multiple modules and updates documentation
* Splits `kda.core` into `kda.calculations`, `kda.diagrams`,
`kda.expressions`, `kda.graphs`, `kda.ode`, and `kda.svd`

* Updates all module imports for testing suite

* Fixes `api.rst` and updates documentation build files for new
module changes
