+++ docs/api.rst
@@ -4,5 +4,10 @@ API Documentation
 .. autosummary::
    :toctree: autosummary
 
-   kda.core
+   kda.calculations
+   kda.diagrams
+   kda.expressions
+   kda.graphs
+   kda.ode
+   kda.svd
    kda.plotting
