--- docs/autosummary/kda.core.rst
@@ -1,23 +1,0 @@
-kda.core
-========
-
-.. automodule:: kda.core
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      BasicMeta
-      Registry
-   
-   
-
-   
-   
-
