+++ docs/autosummary/kda.diagrams.rst
@@ -1,0 +1,33 @@
+﻿kda.diagrams
+============
+
+.. automodule:: kda.diagrams
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      combine
+      generate_all_flux_diagrams
+      generate_directional_partial_diagrams
+      generate_flux_diagrams
+      generate_partial_diagrams
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
