+++ docs/autosummary/kda.ode.rst
@@ -1,0 +1,29 @@
+﻿kda.ode
+=======
+
+.. automodule:: kda.ode
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      ode_solver
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
