+++ docs/autosummary/kda.plotting.rst
@@ -1,4 +1,4 @@
-kda.plotting
+﻿kda.plotting
 ============
 
 .. automodule:: kda.plotting
@@ -9,8 +9,23 @@ kda.plotting
 
    
    
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      draw_ODE_results
+      draw_cycles
+      draw_diagrams
+   
    
 
    
    
    
+
+   
+   
+   
+
+
+
