+++ docs/autosummary/kda.svd.rst
@@ -1,0 +1,29 @@
+﻿kda.svd
+=======
+
+.. automodule:: kda.svd
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      svd_solver
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
