--- docs/getting_started.rst
@@ -1,4 +1,0 @@
-Getting Started
-===============
-
-This page details how to get started with Kinetic Diagram Analysis.
