+++ docs/index.rst
@@ -3,21 +3,47 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to Kinetic Diagram Analysis's documentation!
-=========================================================
+Kinetic Diagram Analysis Documentation
+======================================
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
 
-   getting_started
    api
 
 
+Getting involved
+================
+
+Please report **bugs** or **enhancement requests** through the `Issue
+Tracker <https://github.com/Becksteinlab/kda/issues>`_.
+
+Installing Kinetic Diagram Analysis
+===================================
+
+The easiest approach to install is to use pip_.
+
+Installation with `pip <https://pip.pypa.io/en/latest/>`_ and a
+*minimal set of dependencies*:
+
+.. code-block:: bash
+
+  pip install kda
+
+Source Code
+===========
+
+**Source code** is available from
+https://github.com/Becksteinlab/kda under the `GNU General Public License,
+Version 3 <https://www.gnu.org/licenses/gpl-3.0.en.html>`_. Obtain the sources
+with `git <https://git-scm.com/>`_:
+
+.. code-block:: bash
+
+   git clone https://github.com/Becksteinlab/kda.git
 
 Indices and tables
 ==================
 
 * :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
