+++ kda/tests/test_plotting.py
@@ -10,8 +10,8 @@ from numpy.testing import assert_almost_equal
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 import networkx as nx
-import kda
-import kda.plotting
+
+from kda import plotting, graphs, diagrams, ode
 
 
 @pytest.fixture(scope="module")
@@ -20,16 +20,8 @@ def G4wl():
     k4wl = np.array(
         [[0, k12, 0, k14], [k21, 0, k23, k24], [0, k32, 0, k34], [k41, k42, k43, 0]]
     )
-    k4wls = np.array(
-        [
-            [0, "k12", 0, "k14"],
-            ["k21", 0, "k23", "k24"],
-            [0, "k32", 0, "k34"],
-            ["k41", "k42", "k43", 0],
-        ]
-    )
     G4wl = nx.MultiDiGraph()
-    kda.generate_edges(G4wl, k4wl, k4wls, name_key="name", val_key="val")
+    graphs.generate_edges(G4wl, k4wl)
     return G4wl
 
 
@@ -55,14 +47,6 @@ def results_4wl():
     k4wl = np.array(
         [[0, k12, 0, k14], [k21, 0, k23, k24], [0, k32, 0, k34], [k41, k42, k43, 0]]
     )
-    k4wls = np.array(
-        [
-            [0, "k12", 0, "k14"],
-            ["k21", 0, "k23", "k24"],
-            [0, "k32", 0, "k34"],
-            ["k41", "k42", "k43", 0],
-        ]
-    )
     rate_names4wl = [
         "k12",
         "k21",
@@ -76,33 +60,35 @@ def results_4wl():
         "k42",
     ]
     G4wl = nx.MultiDiGraph()
-    kda.generate_edges(G4wl, k4wl, k4wls, name_key="name", val_key="val")
+    graphs.generate_edges(G4wl, k4wl)
     p4wl = np.array([1, 1, 1, 1]) / 4
-    results4wl = kda.solve_ODE(p4wl, k4wl, t_max=1e2, tol=1e-12, atol=1e-16, rtol=1e-13)
+    results4wl = ode.ode_solver(
+        p4wl, k4wl, t_max=1e2, tol=1e-12, atol=1e-16, rtol=1e-13
+    )
     return results4wl
 
 
 def test_plot_diagram(G4wl, pos_4wl):
-    kda.plotting.draw_diagrams(G4wl)
-    kda.plotting.draw_diagrams(G4wl, pos=pos_4wl)
+    plotting.draw_diagrams(G4wl)
+    plotting.draw_diagrams(G4wl, pos=pos_4wl)
 
 
 def test_plot_cycle(G4wl, pos_4wl):
-    kda.plotting.draw_cycles(G4wl, [0, 1, 3])
-    kda.plotting.draw_cycles(G4wl, [0, 3, 2, 1])
-    kda.plotting.draw_cycles(G4wl, [0, 1, 3], pos=pos_4wl)
-    kda.plotting.draw_cycles(G4wl, [0, 1, 3], pos=pos_4wl, cbt=True)
-    kda.plotting.draw_cycles(G4wl, [[0, 1, 3], [0, 3, 2, 1]], panel=True)
-    kda.plotting.draw_cycles(G4wl, [[0, 1, 3], [0, 3, 2, 1]], panel=True, cbt=True)
-    kda.plotting.draw_cycles(G4wl, [[0, 1, 3], [0, 3, 2, 1]], cbt=True)
+    plotting.draw_cycles(G4wl, [0, 1, 3])
+    plotting.draw_cycles(G4wl, [0, 3, 2, 1])
+    plotting.draw_cycles(G4wl, [0, 1, 3], pos=pos_4wl)
+    plotting.draw_cycles(G4wl, [0, 1, 3], pos=pos_4wl, cbt=True)
+    plotting.draw_cycles(G4wl, [[0, 1, 3], [0, 3, 2, 1]], panel=True)
+    plotting.draw_cycles(G4wl, [[0, 1, 3], [0, 3, 2, 1]], panel=True, cbt=True)
+    plotting.draw_cycles(G4wl, [[0, 1, 3], [0, 3, 2, 1]], cbt=True)
 
 
 def test_plot_diagrams(G4wl, pos_4wl):
-    flux_diags = kda.generate_flux_diagrams(G4wl, [0, 1, 3])
-    kda.plotting.draw_diagrams(flux_diags)
-    kda.plotting.draw_diagrams(flux_diags, pos=pos_4wl, cbt=True)
-    kda.plotting.draw_diagrams(flux_diags, pos=pos_4wl, panel=True, cbt=True)
+    flux_diags = diagrams.generate_flux_diagrams(G4wl, [0, 1, 3])
+    plotting.draw_diagrams(flux_diags)
+    plotting.draw_diagrams(flux_diags, pos=pos_4wl, cbt=True)
+    plotting.draw_diagrams(flux_diags, pos=pos_4wl, panel=True, cbt=True)
 
 
 def test_plot_ODE_probs(results_4wl):
-    kda.plotting.draw_ODE_results(results_4wl)
+    plotting.draw_ODE_results(results_4wl)
